Revert "Added turret"
